Handle globs properly (#638)
