+++ importer-portable/src/main/scala/org/scalablytyped/converter/internal/importer/LibTsSource.scala
@@ -48,6 +48,12 @@ object LibTsSource {
     def fromFileEntry(fromFolder: LibTsSource.FromFolder, files: Option[IArray[String]]): IArray[InFile] =
       files.getOrElse(IArray.Empty).mapNotNone(file => LibraryResolver.file(fromFolder.folder, file))
 
+    def fromFilesGlobEntry(fromFolder: LibTsSource.FromFolder, globs: Option[IArray[String]]): IArray[InFile] = {
+      val baseDirectory = fromFolder.folder.path
+      val files = globs.fold(os.walk(baseDirectory))(GlobWalker.walkFiles(baseDirectory, _)).map(InFile(_))
+      IArray.fromTraversable(files)
+    }
+
     def fromModuleDeclaration(
         fromFolder: LibTsSource.FromFolder,
         files:      Option[Map[String, String]],
@@ -78,19 +84,9 @@ object LibTsSource {
           if (fromTypings.nonEmpty) fromTypings
           else fromModuleDeclaration(f, f.packageJsonOpt.flatMap(_.parsedModules))
 
-        // This used to only find files from the types and typings fields but that is far too limited.
-        // At the very least we need to also include any types specified in the files field.
-        // FIXME: The files field includes file patterns but this assumes that they are all files/folders pointing to
-        //  modules.
-        val filesTypings = fromFileEntry(f, f.packageJsonOpt.flatMap(_.files))
+        val filesTypings = fromFilesGlobEntry(f, f.packageJsonOpt.flatMap(_.files))
 
-        val x = (mainTypings ++ filesTypings).distinct
-
-        println("shortenedFiles:")
-        println(x.mkString("\n"))
-        println
-
-        x
+        (mainTypings ++ filesTypings).distinct
     }
   }
 }
